This creates my first enum and struct.
The enum is for PhysicsBody category and collision bit mask, to allow
me to handle contacts between two physics bodies correctly.

The struct is to create a data type for storing the amount of taps
that someone correctly tapped the ball, and someone missed it in case
I'd want to use that information.

Also moved to instantiating each property in the initializer function init().

Also sped up the game to make it be a little more challenging.

Created the bottomBorder to know when the ball drops to the bottom.

Created scoreLabel to show the current score of the game.

Used the willSet observer to get notified when the TapCount struct is
modified so we can then update the scoreLabel correctly.

Add an extension to GameScene so we can become a contactDelegate
and handle all contact instances.