Improve the mechanism by which we have a 'resting' state of the game.
Previously we were changing whether the ball's physicsBody's dynamic
property was true or not.  This property controls whether it is effected
by the physics simulation in the page.  So, when it's false it doesn't
move at all.

However there was a performance hit when turning dynamic to true as
it required some initial processing to begin emulating how the physics
in the scene should behave.

As a result you could sometimes tap the ball too soon prompting there
to be a noticable lag in the game and make the overall game play not as
smooth as it should be.

Instead I'm taking the route of changing the physicsWorld.speed property
from 0 to the desired value.  This seems to add no performance hit
and accomplishes the same thing we wanted to happen.